requirements.txt
Django==3.2.13
django-import-export==2.7.1
gspread==5.3.2
gunicorn==20.1.0
whitenoise==5.3.0
numpy==1.24.2
oauth2client==4.1.3
argon2-cffi==21.3.0
python-dotenv
